Remove undo, refactor DeckManager and Dialog
